+++ packages/shared/src/graphql/prompt.ts
@@ -15,6 +15,20 @@ export type Prompt = {
   flags?: PromptFlags;
 };
 
+export enum PromptDisplay {
+  TLDR = 'tldr',
+  UpgradeToPlus = 'upgrade-to-plus',
+  ActionableSteps = 'actionable-steps',
+  ChallengeThis = 'challenge-this',
+  CompareAlternatives = 'compare-alternatives',
+  CustomPrompt = 'custom-prompt',
+  ExtractCode = 'extract-code',
+  PracticalExamples = 'practical-examples',
+  RemoveFluff = 'remove-fluff',
+  SimplifyIt = 'simplify-it',
+  SkillsNeeded = 'skills-needed',
+}
+
 export const PROMPTS_QUERY = gql`
   query Prompts {
     prompts {
